✅🚚 Organize most client tests into separate files
Many of the tests that remain in `test_imap` are connection related or
edge cases, e.g: unexpected EOF, STARTTLS stripping, TLS with CA file,
send_command argument translation.  Those will probably stay in this
file indefinitely.  But some of them simply haven't been translated to
`with_fake_server` yet and thus can't easily be moved, e.g: the tests
for idle, disconnect, append, id.  Those will be moved eventually.
