+++ test/net/imap/test_imap.rb
@@ -900,765 +900,6 @@ EOF
     end
   end
 
-  test("#authenticate sends an initial response " \
-       "when supported by both the mechanism and the server") do
-    with_fake_server(
-      preauth: false, cleartext_auth: true, sasl_ir: true
-    ) do |server, imap|
-      imap.authenticate("PLAIN", "test_user", "test-password")
-      cmd = server.commands.pop
-      assert_equal "AUTHENTICATE", cmd.name
-      assert_equal(["PLAIN", ["\x00test_user\x00test-password"].pack("m0")],
-                   cmd.args)
-      assert_empty server.commands
-    end
-  end
-
-  test("#authenticate sends '=' as the initial reponse " \
-       "when the initial response is an empty string") do
-    with_fake_server(
-      preauth: false, cleartext_auth: true,
-      sasl_ir: true, sasl_mechanisms: %i[EXTERNAL],
-    ) do |server, imap|
-      server.on "AUTHENTICATE" do |cmd|
-        server.state.authenticate(server.config.user)
-        cmd.done_ok
-      end
-      imap.authenticate("EXTERNAL")
-      cmd = server.commands.pop
-      assert_equal "AUTHENTICATE", cmd.name
-      assert_equal %w[EXTERNAL =], cmd.args
-      assert_empty server.commands rescue pp server.commands.pop
-    end
-  end
-
-  test("#authenticate never sends an initial response " \
-       "when the server doesn't explicitly support the mechanism") do
-    with_fake_server(
-      preauth: false, cleartext_auth: true,
-      sasl_ir: true, sasl_mechanisms: %i[SCRAM-SHA-1 SCRAM-SHA-256],
-    ) do |server, imap|
-      imap.authenticate("PLAIN", "test_user", "test-password")
-      cmd, cont = 2.times.map { server.commands.pop }
-      assert_equal %w[AUTHENTICATE PLAIN], [cmd.name, *cmd.args]
-      assert_equal(["\x00test_user\x00test-password"].pack("m0"),
-                   cont[:continuation].strip)
-      assert_empty server.commands
-    end
-  end
-
-  test("#authenticate never sends an initial response " \
-       "when the server isn't capable") do
-    with_fake_server(
-      preauth: false, cleartext_auth: true, sasl_ir: false
-    ) do |server, imap|
-      imap.authenticate("PLAIN", "test_user", "test-password")
-      cmd, cont = 2.times.map { server.commands.pop }
-      assert_equal %w[AUTHENTICATE PLAIN], [cmd.name, *cmd.args]
-      assert_equal(["\x00test_user\x00test-password"].pack("m0"),
-                   cont[:continuation].strip)
-      assert_empty server.commands
-    end
-  end
-
-  test("#authenticate never sends an initial response " \
-       "when sasl_ir: false") do
-    [true, false].each do |server_support|
-      with_fake_server(
-        preauth: false, cleartext_auth: true, sasl_ir: server_support
-      ) do |server, imap|
-        imap.authenticate("PLAIN", "test_user", "test-password", sasl_ir: false)
-        cmd, cont = 2.times.map { server.commands.pop }
-        assert_equal %w[AUTHENTICATE PLAIN], [cmd.name, *cmd.args]
-        assert_equal(["\x00test_user\x00test-password"].pack("m0"),
-                     cont[:continuation].strip)
-        assert_empty server.commands
-      end
-    end
-  end
-
-  test("#authenticate never sends an initial response " \
-       "when config.sasl_ir: false") do
-    [true, false].each do |server_support|
-      with_fake_server(
-        preauth: false, cleartext_auth: true, sasl_ir: server_support
-      ) do |server, imap|
-        imap.config.sasl_ir = false
-        imap.authenticate("PLAIN", "test_user", "test-password")
-        cmd, cont = 2.times.map { server.commands.pop }
-        assert_equal %w[AUTHENTICATE PLAIN], [cmd.name, *cmd.args]
-        assert_equal(["\x00test_user\x00test-password"].pack("m0"),
-                     cont[:continuation].strip)
-        assert_empty server.commands
-      end
-    end
-  end
-
-  test("#authenticate never sends an initial response " \
-       "when the mechanism does not support client-first") do
-    with_fake_server(
-      preauth: false, cleartext_auth: true,
-      sasl_ir: true, sasl_mechanisms: %i[DIGEST-MD5]
-    ) do |server, imap|
-      server.on "AUTHENTICATE" do |cmd|
-        response_b64 = cmd.request_continuation(
-          [
-            %w[
-              realm="somerealm"
-              nonce="OA6MG9tEQGm2hh"
-              qop="auth"
-              charset=utf-8
-              algorithm=md5-sess
-            ].join(",")
-          ].pack("m0")
-        )
-        state.commands << {continuation: response_b64}
-        response_b64 = cmd.request_continuation(["rspauth="].pack("m0"))
-        state.commands << {continuation: response_b64}
-        server.state.authenticate(server.config.user)
-        cmd.done_ok
-      end
-      imap.authenticate(:digest_md5, "test_user", "test-password",
-                        warn_deprecation: false)
-      cmd, cont1, cont2 = 3.times.map { server.commands.pop }
-      assert_equal %w[AUTHENTICATE DIGEST-MD5], [cmd.name, *cmd.args]
-      assert_match(%r{\A[a-z0-9+/]+=*\z}i, cont1[:continuation].strip)
-      assert_empty cont2[:continuation].strip
-      assert_empty server.commands
-    end
-  end
-
-  test("#authenticate disconnects and raises SASL::AuthenticationFailed " \
-       "when the server succeeds prematurely") do
-    with_fake_server(
-      preauth: false, cleartext_auth: true,
-      sasl_ir: true, sasl_mechanisms: %i[DIGEST-MD5]
-    ) do |server, imap|
-      server.on "AUTHENTICATE" do |cmd|
-        response_b64 = cmd.request_continuation(
-          [
-            %w[
-              realm="somerealm"
-              nonce="OA6MG9tEQGm2hh"
-              qop="auth"
-              charset=utf-8
-              algorithm=md5-sess
-            ].join(",")
-          ].pack("m0")
-        )
-        state.commands << {continuation: response_b64}
-        server.state.authenticate(server.config.user)
-        cmd.done_ok
-      end
-      assert_raise(Net::IMAP::SASL::AuthenticationIncomplete) do
-        imap.authenticate("DIGEST-MD5", "test_user", "test-password",
-                          warn_deprecation: false)
-      end
-      assert imap.disconnected?
-    end
-  end
-
-  test "#uid_expunge with EXPUNGE responses" do
-    with_fake_server(select: "INBOX",
-                     extensions: %i[UIDPLUS]) do |server, imap|
-      server.on "UID EXPUNGE" do |resp|
-        resp.untagged("1 EXPUNGE")
-        resp.untagged("1 EXPUNGE")
-        resp.untagged("1 EXPUNGE")
-        resp.done_ok
-      end
-      response = imap.uid_expunge(1000..1003)
-      cmd = server.commands.pop
-      assert_equal ["UID EXPUNGE", "1000:1003"], [cmd.name, cmd.args]
-      assert_equal(response, [1, 1, 1])
-    end
-  end
-
-  test "#uid_expunge with VANISHED response" do
-    with_fake_server(select: "INBOX",
-                     extensions: %i[UIDPLUS]) do |server, imap|
-      server.on "UID EXPUNGE" do |resp|
-        resp.untagged("VANISHED 1001,1003")
-        resp.done_ok
-      end
-      response = imap.uid_expunge(1000..1003)
-      cmd = server.commands.pop
-      assert_equal ["UID EXPUNGE", "1000:1003"], [cmd.name, cmd.args]
-      assert_equal(
-        Net::IMAP::VanishedData[uids: [1001, 1003], earlier: false],
-        response
-      )
-      assert_equal([], imap.clear_responses("VANISHED"))
-    end
-  end
-
-  def test_uidplus_appenduid
-    with_fake_server(select: "INBOX",
-                     extensions: %i[UIDPLUS]) do |server, imap|
-      server.on "APPEND" do |cmd|
-        cmd.done_ok code: "APPENDUID 38505 3955"
-      end
-      resp = imap.append("inbox", <<~EOF.gsub(/\n/, "\r\n"), [:Seen], Time.now)
-        Subject: hello
-        From: shugo@ruby-lang.org
-        To: shugo@ruby-lang.org
-
-        hello world
-      EOF
-      assert_equal([38505, nil, [3955]], resp.data.code.data.to_a)
-      assert_equal "APPEND", server.commands.pop.name
-    end
-  end
-
-  def test_uidplus_copyuid_multiple
-    with_fake_server(select: "INBOX",
-                     extensions: %i[UIDPLUS]) do |server, imap|
-      server.on "UID COPY" do |cmd|
-        cmd.done_ok code: "COPYUID 38505 3955,3960:3962 3963:3966"
-      end
-      resp = imap.uid_copy([3955,3960..3962], 'trash')
-      cmd  = server.commands.pop
-      assert_equal(["UID COPY", "3955,3960:3962 trash"], [cmd.name, cmd.args])
-      assert_equal(
-        [38505, [3955, 3960, 3961, 3962], [3963, 3964, 3965, 3966]],
-        resp.data.code.data.to_a
-      )
-    end
-  end
-
-  def test_uidplus_copyuid_single
-    with_fake_server(select: "INBOX",
-                     extensions: %i[UIDPLUS]) do |server, imap|
-      server.on "UID COPY" do |cmd|
-        cmd.done_ok code: "COPYUID 38505 3955 3967"
-      end
-      resp = imap.uid_copy(3955, 'trash')
-      cmd  = server.commands.pop
-      assert_equal(["UID COPY", "3955 trash"], [cmd.name, cmd.args])
-      assert_equal([38505, [3955], [3967]], resp.data.code.data.to_a)
-    end
-  end
-
-  def test_uidplus_uidnotsticky
-    with_fake_server(extensions: %i[UIDPLUS]) do |server, imap|
-      server.config.mailboxes["trash"] = { uidnotsticky: true }
-      imap.select('trash')
-      assert imap.responses("NO", &:to_a).any? {
-        _1.code == Net::IMAP::ResponseCode.new('UIDNOTSTICKY', nil)
-      }
-    end
-  end
-
-  def test_enable
-    with_fake_server(
-      with_extensions: %i[ENABLE CONDSTORE UTF8=ACCEPT],
-      capabilities_enablable: %w[CONDSTORE UTF8=ACCEPT]
-    ) do |server, imap|
-      cmdq = server.commands
-
-      result1 = imap.enable(%w[CONDSTORE x-pig-latin])
-      result2 = imap.enable(:utf8, "condstore QResync")
-      result3 = imap.enable(:utf8, "UTF8=ACCEPT", "UTF8=ONLY")
-      cmd1, cmd2, cmd3 = Array.new(3) { cmdq.pop.raw.strip }
-
-      assert_equal "RUBY0001 ENABLE CONDSTORE x-pig-latin",         cmd1
-      assert_equal "RUBY0002 ENABLE UTF8=ACCEPT condstore QResync", cmd2
-      assert_equal "RUBY0003 ENABLE UTF8=ACCEPT",                   cmd3
-      assert_empty cmdq
-
-      assert_equal %w[CONDSTORE],   result1
-      assert_equal %w[UTF8=ACCEPT], result2
-      assert_equal [],              result3
-    end
-  end
-
-  test "#select with condstore" do
-    with_fake_server do |server, imap|
-      imap.select "inbox", condstore: true
-      assert_equal("RUBY0001 SELECT inbox (CONDSTORE)",
-                   server.commands.pop.raw.strip)
-    end
-  end
-
-  test "#examine with condstore" do
-    with_fake_server do |server, imap|
-      imap.examine "inbox", condstore: true
-      assert_equal("RUBY0001 EXAMINE inbox (CONDSTORE)",
-                   server.commands.pop.raw.strip)
-    end
-  end
-
-  test "#fetch with FETCH responses" do
-    with_fake_server select: "inbox" do |server, imap|
-      server.on("FETCH") do |resp|
-        resp.untagged("123 FETCH (UID 1111 FLAGS (\\Seen $MDNSent))")
-        resp.untagged("456 FETCH (UID 4444 FLAGS (\\Seen \\Answered))")
-        resp.untagged("789 FETCH (UID 7777 FLAGS ())")
-        resp.done_ok
-      end
-      fetched = imap.fetch [123, 456, 789], %w[UID FLAGS]
-      assert_equal 3, fetched.size
-      assert_instance_of Net::IMAP::FetchData, fetched[0]
-      assert_instance_of Net::IMAP::FetchData, fetched[1]
-      assert_instance_of Net::IMAP::FetchData, fetched[2]
-      assert_equal 123,  fetched[0].seqno
-      assert_equal 456,  fetched[1].seqno
-      assert_equal 789,  fetched[2].seqno
-      assert_equal 1111, fetched[0].uid
-      assert_equal 4444, fetched[1].uid
-      assert_equal 7777, fetched[2].uid
-      assert_equal [:Seen, "$MDNSent"], fetched[0].flags
-      assert_equal [:Seen, :Answered],  fetched[1].flags
-      assert_equal [],                  fetched[2].flags
-      assert_equal("RUBY0002 FETCH 123,456,789 (UID FLAGS)",
-                   server.commands.pop.raw.strip)
-    end
-  end
-
-  test "#uid_fetch with UIDFETCH responses" do
-    with_fake_server select: "inbox" do |server, imap|
-      server.on("UID FETCH") do |resp|
-        resp.untagged("1111 UIDFETCH (FLAGS (\\Seen $MDNSent))")
-        resp.untagged("4444 UIDFETCH (FLAGS (\\Seen \\Answered))")
-        resp.untagged("7777 UIDFETCH (FLAGS ())")
-        resp.done_ok
-      end
-      fetched = imap.uid_fetch [123, 456, 789], %w[FLAGS]
-      assert_equal 3, fetched.size
-      assert_instance_of Net::IMAP::UIDFetchData, fetched[0]
-      assert_instance_of Net::IMAP::UIDFetchData, fetched[1]
-      assert_instance_of Net::IMAP::UIDFetchData, fetched[2]
-      assert_equal 1111, fetched[0].uid
-      assert_equal 4444, fetched[1].uid
-      assert_equal 7777, fetched[2].uid
-      assert_equal [:Seen, "$MDNSent"], fetched[0].flags
-      assert_equal [:Seen, :Answered],  fetched[1].flags
-      assert_equal [],                  fetched[2].flags
-      assert_equal("RUBY0002 UID FETCH 123,456,789 (FLAGS)",
-                   server.commands.pop.raw.strip)
-    end
-  end
-
-  test "#fetch with changedsince" do
-    with_fake_server select: "inbox" do |server, imap|
-      server.on("FETCH", &:done_ok)
-      fetched = imap.fetch 1..-1, %w[FLAGS], changedsince: 12345
-      assert_empty fetched
-      assert_equal("RUBY0002 FETCH 1:* (FLAGS) (CHANGEDSINCE 12345)",
-                   server.commands.pop.raw.strip)
-    end
-  end
-
-  test "#uid_fetch with changedsince" do
-    with_fake_server select: "inbox" do |server, imap|
-      server.on("UID FETCH", &:done_ok)
-      fetched = imap.uid_fetch 1..-1, %w[FLAGS], changedsince: 12345
-      assert_empty fetched
-      assert_equal("RUBY0002 UID FETCH 1:* (FLAGS) (CHANGEDSINCE 12345)",
-                   server.commands.pop.raw.strip)
-    end
-  end
-
-  test "#uid_fetch with partial" do
-    with_fake_server select: "inbox" do |server, imap|
-      server.on("UID FETCH", &:done_ok)
-      imap.uid_fetch 1.., "FAST", partial: 1..500
-      assert_equal("RUBY0002 UID FETCH 1:* FAST (PARTIAL 1:500)",
-                   server.commands.pop.raw.strip)
-      imap.uid_fetch 1.., "FAST", partial: 1...501
-      assert_equal("RUBY0003 UID FETCH 1:* FAST (PARTIAL 1:500)",
-                   server.commands.pop.raw.strip)
-      imap.uid_fetch 1.., "FAST", partial: -500..-1
-      assert_equal("RUBY0004 UID FETCH 1:* FAST (PARTIAL -500:-1)",
-                   server.commands.pop.raw.strip)
-      imap.uid_fetch 1.., "FAST", partial: -500...-1
-      assert_equal("RUBY0005 UID FETCH 1:* FAST (PARTIAL -500:-2)",
-                   server.commands.pop.raw.strip)
-      imap.uid_fetch 1.., "FAST", partial: 1..20, changedsince: 1234
-      assert_equal("RUBY0006 UID FETCH 1:* FAST (PARTIAL 1:20 CHANGEDSINCE 1234)",
-                   server.commands.pop.raw.strip)
-    end
-  end
-
-  test "#store with FETCH responses" do
-    with_fake_server select: "inbox" do |server, imap|
-      server.on("STORE") do |resp|
-        resp.untagged("123 FETCH (UID 1111 FLAGS (\\Seen $MDNSent))")
-        resp.untagged("456 FETCH (UID 4444 FLAGS (\\Seen \\Answered))")
-        resp.untagged("789 FETCH (UID 7777 FLAGS (\\Seen))")
-        resp.done_ok
-      end
-      changed = imap.store [123, 456, 789], "+FLAGS", %i[Seen]
-      assert_equal("RUBY0002 STORE 123,456,789 +FLAGS (\\Seen)",
-                   server.commands.pop.raw.strip)
-      assert_equal 3, changed.size
-      assert_instance_of Net::IMAP::FetchData, changed[0]
-      assert_instance_of Net::IMAP::FetchData, changed[1]
-      assert_instance_of Net::IMAP::FetchData, changed[2]
-      assert_equal 123,  changed[0].seqno
-      assert_equal 456,  changed[1].seqno
-      assert_equal 789,  changed[2].seqno
-      assert_equal 1111, changed[0].uid
-      assert_equal 4444, changed[1].uid
-      assert_equal 7777, changed[2].uid
-      assert_equal [:Seen, "$MDNSent"], changed[0].flags
-      assert_equal [:Seen, :Answered],  changed[1].flags
-      assert_equal [:Seen],             changed[2].flags
-    end
-  end
-
-  test "#uid_store with UIDFETCH responses" do
-    with_fake_server select: "inbox" do |server, imap|
-      server.on("UID STORE") do |resp|
-        resp.untagged("1111 UIDFETCH (FLAGS (\\Seen $MDNSent))")
-        resp.untagged("4444 UIDFETCH (FLAGS (\\Seen \\Answered))")
-        resp.untagged("7777 UIDFETCH (FLAGS (\\Seen))")
-        resp.done_ok
-      end
-      changed = imap.uid_store [123, 456, 789], "+FLAGS", %i[Seen]
-      assert_equal("RUBY0002 UID STORE 123,456,789 +FLAGS (\\Seen)",
-                   server.commands.pop.raw.strip)
-      assert_equal 3, changed.size
-      assert_instance_of Net::IMAP::UIDFetchData, changed[0]
-      assert_instance_of Net::IMAP::UIDFetchData, changed[1]
-      assert_instance_of Net::IMAP::UIDFetchData, changed[2]
-      assert_equal 1111, changed[0].uid
-      assert_equal 4444, changed[1].uid
-      assert_equal 7777, changed[2].uid
-      assert_equal [:Seen, "$MDNSent"], changed[0].flags
-      assert_equal [:Seen, :Answered],  changed[1].flags
-      assert_equal [:Seen],             changed[2].flags
-    end
-  end
-
-  test "#store with unchangedsince" do
-    with_fake_server select: "inbox" do |server, imap|
-      server.on("STORE", &:done_ok)
-      imap.store 1..-1, "FLAGS", %i[Deleted], unchangedsince: 12345
-      assert_equal(
-        "RUBY0002 STORE 1:* (UNCHANGEDSINCE 12345) FLAGS (\\Deleted)",
-        server.commands.pop.raw.strip
-      )
-    end
-  end
-
-  test "#uid_store with changedsince" do
-    with_fake_server select: "inbox" do |server, imap|
-      server.on("UID STORE", &:done_ok)
-      imap.uid_store 1..-1, "FLAGS", %i[Deleted], unchangedsince: 987
-      assert_equal(
-        "RUBY0002 UID STORE 1:* (UNCHANGEDSINCE 987) FLAGS (\\Deleted)",
-        server.commands.pop.raw.strip
-      )
-    end
-  end
-
-  test "#expunge with EXPUNGE responses" do
-    with_fake_server(select: "INBOX") do |server, imap|
-      server.on "EXPUNGE" do |resp|
-        resp.untagged("1 EXPUNGE")
-        resp.untagged("1 EXPUNGE")
-        resp.untagged("99 EXPUNGE")
-        resp.done_ok
-      end
-      response = imap.expunge
-      cmd = server.commands.pop
-      assert_equal ["EXPUNGE", nil], [cmd.name, cmd.args]
-      assert_equal [1, 1, 99], response
-      assert_equal [], imap.clear_responses("EXPUNGED")
-    end
-  end
-
-  test "#expunge with a VANISHED response" do
-    with_fake_server(select: "INBOX") do |server, imap|
-      server.on "EXPUNGE" do |resp|
-        resp.untagged("VANISHED 15:456")
-        resp.done_ok
-      end
-      response = imap.expunge
-      cmd = server.commands.pop
-      assert_equal ["EXPUNGE", nil], [cmd.name, cmd.args]
-      assert_equal(
-        Net::IMAP::VanishedData[uids: [15..456], earlier: false],
-        response
-      )
-      assert_equal([], imap.clear_responses("VANISHED"))
-    end
-  end
-
-  test "#expunge with multiple VANISHED responses" do
-    with_fake_server(select: "INBOX") do |server, imap|
-      server.unsolicited("VANISHED 86")
-      server.on "EXPUNGE" do |resp|
-        resp.untagged("VANISHED (EARLIER) 1:5,99,123")
-        resp.untagged("VANISHED 15,456")
-        resp.untagged("VANISHED (EARLIER) 987,1001")
-        resp.done_ok
-      end
-      response = imap.expunge
-      cmd = server.commands.pop
-      assert_equal ["EXPUNGE", nil], [cmd.name, cmd.args]
-      assert_equal(
-        Net::IMAP::VanishedData[uids: [15, 86, 456], earlier: false],
-        response
-      )
-      assert_equal(
-        [
-          Net::IMAP::VanishedData[uids: [1..5, 99, 123], earlier: true],
-          Net::IMAP::VanishedData[uids: [987, 1001],     earlier: true],
-        ],
-        imap.clear_responses("VANISHED")
-      )
-    end
-  end
-
-  def test_close
-    with_fake_server(select: "inbox") do |server, imap|
-      resp = imap.close
-      assert_equal("RUBY0002 CLOSE", server.commands.pop.raw.strip)
-      assert_equal([Net::IMAP::TaggedResponse, "RUBY0002", "OK"],
-                   [resp.class, resp.tag, resp.name])
-      assert_empty server.commands
-    end
-  end
-
-  def test_unselect
-    with_fake_server(select: "inbox") do |server, imap|
-      resp = imap.unselect
-      sent = server.commands.pop
-      assert_equal(["UNSELECT", nil], [sent.name, sent.args])
-      assert_equal([Net::IMAP::TaggedResponse, "RUBY0002", "OK"],
-                   [resp.class, resp.tag, resp.name])
-      assert_empty server.commands
-    end
-  end
-
-  test("missing server ENABLED response") do
-    with_fake_server do |server, imap|
-      server.on "ENABLE", &:done_ok
-      enabled = imap.enable "foo", "bar", "baz"
-      assert_equal [], enabled
-    end
-  end
-
-  test("#search/#uid_search") do
-    with_fake_server do |server, imap|
-      search_result = Net::IMAP::SearchResult[
-        1, 2, 3, 5, 8, 13, 21, 34, 55, modseq: 1234
-      ]
-      search_resp = ->cmd do
-        cmd.puts search_result.to_s("SEARCH")
-        cmd.done_ok
-      end
-
-      server.on "SEARCH", &search_resp
-      assert_equal search_result, imap.search(["subject", "hello world",
-                                               [1..5, 8, 10..-1]])
-      cmd = server.commands.pop
-      assert_equal(
-        ["SEARCH",'subject "hello world" 1:5,8,10:*'],
-        [cmd.name, cmd.args]
-      )
-
-      imap.search(["OR", 1..1000, -1, "UID", 12345..-1])
-      assert_equal "OR 1:1000 * UID 12345:*", server.commands.pop.args
-
-      imap.search([1..1000, "UID", 12345..])
-      assert_equal "1:1000 UID 12345:*", server.commands.pop.args
-
-      # Unfortunately, we can't send every sequence-set string directly
-      imap.search(["SUBJECT", "1,*"])
-      assert_equal 'SUBJECT "1,*"', server.commands.pop.args
-
-      imap.search(["subject", "hello", Set[1, 2, 3, 4, 5, 8, *(10..100)]])
-      assert_equal "subject hello 1:5,8,10:100", server.commands.pop.args
-
-      imap.search('SUBJECT "Hello world"', "UTF-8")
-      assert_equal 'CHARSET UTF-8 SUBJECT "Hello world"', server.commands.pop.args
-
-      imap.search('CHARSET UTF-8 SUBJECT "Hello world"')
-      assert_equal 'CHARSET UTF-8 SUBJECT "Hello world"', server.commands.pop.args
-
-      imap.search('SUBJECT "Hello world"', charset: "UTF-8")
-      assert_equal 'CHARSET UTF-8 SUBJECT "Hello world"', server.commands.pop.args
-
-      imap.search([:*])
-      assert_equal "*", server.commands.pop.args
-
-      seqset_coercible = Object.new
-      def seqset_coercible.to_sequence_set
-        Net::IMAP::SequenceSet[1..9]
-      end
-      imap.search([seqset_coercible])
-      assert_equal "1:9", server.commands.pop.args
-
-      server.on "UID SEARCH", &search_resp
-      assert_equal search_result, imap.uid_search(["subject", "hello",
-                                                   [1..22, 30..-1]])
-      cmd = server.commands.pop
-      assert_equal ["UID SEARCH", "subject hello 1:22,30:*"], [cmd.name, cmd.args]
-
-      assert_equal search_result, imap.search(
-        "RETURN (COUNT) NOT (FLAGGED (OR SEEN ANSWERED))"
-      )
-      cmd = server.commands.pop
-      assert_equal "RETURN (COUNT) NOT (FLAGGED (OR SEEN ANSWERED))", cmd.args
-
-      assert_equal search_result, imap.search([
-        "RETURN", %w(MIN MAX COUNT), "NOT", ["FLAGGED", %w(OR SEEN ANSWERED)]
-      ])
-      cmd = server.commands.pop
-      assert_equal "RETURN (MIN MAX COUNT) NOT (FLAGGED (OR SEEN ANSWERED))", cmd.args
-
-      assert_equal search_result, imap.search(
-        ["NOT", ["FLAGGED", %w(OR SEEN ANSWERED)]], return: %w(MIN MAX COUNT)
-      )
-      cmd = server.commands.pop
-      assert_equal "RETURN (MIN MAX COUNT) NOT (FLAGGED (OR SEEN ANSWERED))", cmd.args
-
-      assert_equal search_result, imap.search(
-        ["UID", 1234..], return: %w(PARTIAL -500:-1)
-      )
-      cmd = server.commands.pop
-      assert_equal "RETURN (PARTIAL -500:-1) UID 1234:*", cmd.args
-
-      assert_equal search_result, imap.search(
-        ["UID", 1234..], return: [:PARTIAL, "-500:-1"]
-      )
-      cmd = server.commands.pop
-      assert_equal "RETURN (PARTIAL -500:-1) UID 1234:*", cmd.args
-
-      assert_equal search_result, imap.search(
-        ["UID", 1234..], return: [:PARTIAL, -500..-1, :FOO, 1..]
-      )
-      cmd = server.commands.pop
-      assert_equal "RETURN (PARTIAL -500:-1 FOO 1:*) UID 1234:*", cmd.args
-    end
-  end
-
-  test("#search/#uid_search with invalid arguments") do
-    with_fake_server do |server, imap|
-      server.on "SEARCH"     do |cmd| cmd.fail_no "should fail before this" end
-      server.on "UID SEARCH" do |cmd| cmd.fail_no "should fail before this" end
-
-      assert_raise(ArgumentError) do
-        imap.search(["charset", "foo", "ALL"], "bar")
-      end
-      assert_raise(ArgumentError) do
-        imap.search("charset foo ALL", "bar")
-      end
-      assert_raise(ArgumentError) do
-        imap.search(["ALL"], "foo", charset: "bar")
-      end
-      assert_raise(ArgumentError) do
-        imap.search(["charset", "foo", "ALL"], charset: "bar")
-      end
-      assert_raise(ArgumentError) do
-        imap.search("charset foo ALL", charset: "bar")
-      end
-      # Parsing return opts is too complicated, for now.
-      # assert_raise(ArgumentError) do
-      #   imap.search("return () charset foo ALL", "bar")
-      # end
-
-      assert_raise(ArgumentError) do
-        imap.search(["retURN", %w(foo bar), "ALL"], return: %w[foo bar])
-      end
-      assert_raise(ArgumentError) do
-        imap.search("RETURN (foo bar) ALL", return: %w[foo bar])
-      end
-      assert_raise(TypeError) do
-        imap.search("ALL", return: "foo bar")
-      end
-      assert_raise(TypeError) do
-        imap.search(["retURN", "foo bar", "ALL"])
-      end
-    end
-  end
-
-  test("#search/#uid_search with ESEARCH or IMAP4rev2") do
-    with_fake_server do |server, imap|
-      # Example from RFC9051, 6.4.4:
-      #   C: A282 SEARCH RETURN (MIN COUNT) FLAGGED
-      #       SINCE 1-Feb-1994 NOT FROM "Smith"
-      #   S: * ESEARCH (TAG "A282") MIN 2 COUNT 3
-      #   S: A282 OK SEARCH completed
-      server.on "SEARCH" do |cmd|
-        cmd.untagged "ESEARCH", "(TAG \"unrelated1\") MIN 1 COUNT 2"
-        cmd.untagged "ESEARCH", "(TAG %p) MIN 2 COUNT 3" % [cmd.tag]
-        cmd.untagged "ESEARCH", "(TAG \"unrelated2\") MIN 222 COUNT 333"
-        cmd.done_ok
-      end
-      result = imap.search(
-        'RETURN (MIN COUNT) FLAGGED SINCE 1-Feb-1994 NOT FROM "Smith"'
-      )
-      cmd = server.commands.pop
-      assert_equal Net::IMAP::ESearchResult.new(
-        cmd.tag, false, [["MIN", 2], ["COUNT", 3]]
-      ), result
-      esearch_responses = imap.clear_responses("ESEARCH")
-      assert_equal 2, esearch_responses.count
-      refute esearch_responses.include?(result)
-    end
-  end
-
-  test("missing server ESEARCH response") do
-    with_fake_server do |server, imap|
-      # Example from RFC9051, 6.4.4:
-      #   C: A282 SEARCH RETURN (SAVE) FLAGGED SINCE 1-Feb-1994 NOT FROM "Smith"
-      #   S: A282 OK SEARCH completed, result saved
-      server.on "SEARCH"     do |cmd| cmd.done_ok "result saved" end
-      server.on "UID SEARCH" do |cmd| cmd.done_ok "result saved" end
-      result = imap.search(
-        'RETURN (SAVE) FLAGGED SINCE 1-Feb-1994 NOT FROM "Smith"'
-      )
-      assert_pattern do
-        result => Net::IMAP::ESearchResult[uid: false, tag: /^RUBY\d+/, data: []]
-      end
-      result = imap.uid_search(
-        'RETURN (SAVE) FLAGGED SINCE 1-Feb-1994 NOT FROM "Smith"'
-      )
-      assert_pattern do
-        result => Net::IMAP::ESearchResult[uid: true, tag: /^RUBY\d+/, data: []]
-      end
-    end
-  end
-
-  test("missing server SEARCH response") do
-    with_fake_server do |server, imap|
-      server.on "SEARCH",     &:done_ok
-      server.on "UID SEARCH", &:done_ok
-      found = imap.search ["subject", "hello"]
-      assert_instance_of Net::IMAP::SearchResult, found
-      assert_empty found
-      found = imap.uid_search ["subject", "hello"]
-      assert_instance_of Net::IMAP::SearchResult, found
-      assert_empty found
-    end
-  end
-
-  test("missing server SORT response") do
-    with_fake_server do |server, imap|
-      server.on "SORT",       &:done_ok
-      server.on "UID SORT",   &:done_ok
-      found = imap.sort ["INTERNALDATE"], ["subject", "hello"], "UTF-8"
-      assert_equal [], found
-      found = imap.uid_sort ["INTERNALDATE"], ["subject", "hello"], "UTF-8"
-      assert_equal [], found
-    end
-  end
-
-  test("missing server THREAD response") do
-    with_fake_server do |server, imap|
-      server.on "THREAD",     &:done_ok
-      server.on "UID THREAD", &:done_ok
-      found = imap.thread "REFERENCES", ["subject", "hello"], "UTF-8"
-      assert_equal [], found
-      found = imap.uid_thread "REFERENCES", ["subject", "hello"], "UTF-8"
-      assert_equal [], found
-    end
-  end
-
   private
 
   def imaps_test(timeout: 10)
